requirements.txt
torch
torchvision
timm
streamlit
numpy
pillow
scikit-learn
matplotlib
seaborn
opencv-python
